+++ lulzlyn/lulzlyn/utilities.hpp
@@ -5,6 +5,6 @@
 
 namespace utilities
 {
-	std::uintptr_t* pattern_scan(const char* module_name, const char* signature);
+	std::uint8_t* pattern_scan(const char* module_name, const char* signature);
 	HBITMAP load_bitmap();
 }
